+++ src/google_music_scripts/commands/delete.py
@@ -5,12 +5,12 @@ import google_music
 from logzero import logger
 
 from google_music_scripts.__about__ import __title__, __version__
-from google_music_scripts.cli import split_filter_strings
+from google_music_scripts.cli import CONTEXT_SETTINGS, parse_filters
 from google_music_scripts.config import configure_logging
 from google_music_scripts.core import filter_songs
 
 
-@click.command()
+@click.command(context_settings=CONTEXT_SETTINGS)
 @click.version_option(__version__, '-V', '--version', prog_name=__title__, message="%(prog)s %(version)s")
 @click.option('-l', '--log', is_flag=True, default=False, help="Log to file.")
 @click.option('-v', '--verbose', count=True)
@@ -21,27 +21,13 @@ from google_music_scripts.core import filter_songs
 	help="Your Google username or e-mail address.\nUsed to separate saved credentials."
 )
 @click.option('--device-id', metavar='ID', help="A mobile device id.")
-@click.option(
-	'-f', '--include-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-fa', '--all-includes', is_flag=True, default=False,
-	help="Songs must match all include filter criteria to be included."
-)
-@click.option(
-	'-F', '--exclude-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-Fa', '--all-excludes', is_flag=True, default=False,
-	help="Songs must match all exclude filter criteria to be included."
-)
 @click.option('-y', '--yes', is_flag=True, default=False, help="Delete songs without asking for confirmation.")
+@click.option(
+	'-f', '--filters', metavar='FILTER', multiple=True,
+	callback=parse_filters, help="Metadata filters."
+)
 def delete(
-	log, verbose, quiet, dry_run, username, device_id,
-	include_filter, all_includes, exclude_filter, all_excludes, yes
-):
+	log, verbose, quiet, dry_run, username, device_id, yes, filters):
 	"""Delete songs from a Google Music library."""
 
 	configure_logging(verbose - quiet, log_to_file=log)
@@ -52,11 +38,7 @@ def delete(
 	if not mc.is_authenticated:
 		sys.exit("Failed to authenticate client.")
 
-	to_delete = filter_songs(
-		mc.songs(),
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
+	to_delete = filter_songs(mc.songs(), filters)
 
 	if not to_delete:
 		logger.info("No songs to delete")