+++ src/google_music_scripts/commands/download.py
@@ -6,12 +6,12 @@ import google_music
 from logzero import logger
 
 from google_music_scripts.__about__ import __title__, __version__
-from google_music_scripts.cli import CustomPath, split_filter_strings
+from google_music_scripts.cli import CONTEXT_SETTINGS, CustomPath, parse_filters
 from google_music_scripts.config import configure_logging
 from google_music_scripts.core import download_songs, filter_songs
 
 
-@click.command()
+@click.command(context_settings=CONTEXT_SETTINGS)
 @click.version_option(__version__, '-V', '--version', prog_name=__title__, message="%(prog)s %(version)s")
 @click.option('-l', '--log', is_flag=True, default=False, help="Log to file.")
 @click.option('-v', '--verbose', count=True)
@@ -26,28 +26,15 @@ from google_music_scripts.core import download_songs, filter_songs
 	help="A unique id given as a MAC address (e.g. '00:11:22:33:AA:BB').\nThis should only be provided when the default does not work."
 )
 @click.option(
-	'-f', '--include-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-fa', '--all-includes', is_flag=True, default=False,
-	help="Songs must match all include filter criteria to be included."
-)
-@click.option(
-	'-F', '--exclude-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-Fa', '--all-excludes', is_flag=True, default=False,
-	help="Songs must match all exclude filter criteria to be included."
-)
-@click.option(
 	'-o', '--output', metavar='TEMPLATE_PATH', default=os.getcwd(), type=CustomPath(),
 	help="Output file or directory name which can include template patterns."
 )
+@click.option(
+	'-f', '--filters', metavar='FILTER', multiple=True,
+	callback=parse_filters, help="Metadata filters."
+)
 def download(
-	log, verbose, quiet, dry_run, username, uploader_id,
-	include_filter, all_includes, exclude_filter, all_excludes, output):
+	log, verbose, quiet, dry_run, username, uploader_id, output, filters):
 	"""Download songs from a Google Music library."""
 
 	configure_logging(verbose - quiet, log_to_file=log)
@@ -58,12 +45,7 @@ def download(
 	if not mm.is_authenticated:
 		sys.exit("Failed to authenticate client.")
 
-	to_download = filter_songs(
-		mm.songs(),
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
-
+	to_download = filter_songs(mm.songs(), filters)
 	to_download.sort(key=lambda song: (song.get('artist'), song.get('album'), song.get('track_number')))
 
 	if not to_download: