+++ src/google_music_scripts/commands/sync.py
@@ -8,7 +8,7 @@ from click_default_group import DefaultGroup
 from logzero import logger
 
 from google_music_scripts.__about__ import __title__, __version__
-from google_music_scripts.cli import CustomPath, default_to_cwd, split_filter_strings
+from google_music_scripts.cli import CONTEXT_SETTINGS, CustomPath, default_to_cwd, parse_filters
 from google_music_scripts.config import configure_logging
 from google_music_scripts.core import (
 	download_songs, filter_songs, get_local_songs, upload_songs
@@ -16,7 +16,7 @@ from google_music_scripts.core import (
 from google_music_scripts.utils import template_to_base_path
 
 
-@click.group(cls=DefaultGroup, default='up', default_if_no_args=True)
+@click.group(cls=DefaultGroup, default='up', default_if_no_args=True, context_settings=CONTEXT_SETTINGS)
 @click.version_option(__version__, '-V', '--version', prog_name=__title__, message="%(prog)s %(version)s")
 def sync():
 	"""Sync songs to/from a Google Music library."""
@@ -47,29 +47,17 @@ def sync():
 	help="Set maximum depth of recursion when scanning for local files.\nDefault is infinite recursion."
 )
 @click.option(
-	'-f', '--include-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-fa', '--all-includes', is_flag=True, default=False,
-	help="Songs must match all include filter criteria to be included."
-)
-@click.option(
-	'-F', '--exclude-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-Fa', '--all-excludes', is_flag=True, default=False,
-	help="Songs must match all exclude filter criteria to be included."
-)
-@click.option(
 	'-o', '--output', metavar='TEMPLATE_PATH', default=os.getcwd(), type=CustomPath(),
 	help="Output file or directory name which can include template patterns."
 )
+@click.option(
+	'-f', '--filters', metavar='FILTER', multiple=True,
+	callback=parse_filters, help="Metadata filters."
+)
 @click.argument('include-paths', nargs=-1, type=CustomPath(resolve_path=True))
 def sync_down(
-	log, verbose, quiet, dry_run, username, uploader_id, no_recursion, max_depth,
-	include_filter, all_includes, exclude_filter, all_excludes, output, include_paths):
+	log, verbose, quiet, dry_run, username, uploader_id,
+	no_recursion, max_depth, output, filters, include_paths):
 	"""Sync songs from a Google Music library."""
 
 	configure_logging(verbose - quiet, log_to_file=log)
@@ -85,23 +73,14 @@ def sync_down(
 	elif max_depth is None:
 		max_depth = float('inf')
 
-	google_songs = filter_songs(
-		mm.songs(),
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
+	google_songs = filter_songs(mm.songs(), filters)
 
 	base_path = template_to_base_path(output, google_songs)
 	filepaths = [base_path]
-
 	if include_paths:
 		filepaths.extend(include_paths)
 
-	local_songs = get_local_songs(
-		filepaths, max_depth=max_depth,
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
+	local_songs = get_local_songs(filepaths, max_depth=max_depth)
 
 	logger.info("Comparing song collections")
 	to_download = list(
@@ -153,22 +132,6 @@ def sync_down(
 	'--max-depth', metavar='DEPTH', type=int,
 	help="Set maximum depth of recursion when scanning for local files.\nDefault is infinite recursion."
 )
-@click.option(
-	'-f', '--include-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-fa', '--all-includes', is_flag=True, default=False,
-	help="Songs must match all include filter criteria to be included."
-)
-@click.option(
-	'-F', '--exclude-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-Fa', '--all-excludes', is_flag=True, default=False,
-	help="Songs must match all exclude filter criteria to be included."
-)
 @click.option('--delete-on-success', is_flag=True, default=False, help="Delete successfully uploaded local files.")
 @click.option(
 	'--transcode-lossless/--no-transcode-lossless', is_flag=True, default=True,
@@ -178,10 +141,14 @@ def sync_down(
 	'--transcode-lossy/--no-transcode-lossy', is_flag=True, default=True,
 	help="Transcode non-MP3 lossy files to MP3 for upload."
 )
+@click.option(
+	'-f', '--filters', metavar='FILTER', multiple=True,
+	callback=parse_filters, help="Metadata filters."
+)
 @click.argument('input-paths', nargs=-1, type=CustomPath(resolve_path=True), callback=default_to_cwd)
 def sync_up(
-	log, verbose, quiet, dry_run, username, uploader_id, no_recursion, max_depth, delete_on_success,
-	transcode_lossless, transcode_lossy, include_filter, all_includes, exclude_filter, all_excludes, input_paths):
+	log, verbose, quiet, dry_run, username, uploader_id, no_recursion, max_depth,
+	delete_on_success, transcode_lossless, transcode_lossy, filters, input_paths):
 	"""Sync songs to a Google Music library."""
 
 	configure_logging(verbose - quiet, log_to_file=log)
@@ -197,17 +164,8 @@ def sync_up(
 	elif max_depth is None:
 		max_depth = float('inf')
 
-	google_songs = filter_songs(
-		mm.songs(),
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
-
-	local_songs = get_local_songs(
-		input_paths, max_depth=max_depth,
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
+	google_songs = mm.songs()
+	local_songs = get_local_songs(input_paths, filters=filters, max_depth=max_depth)
 
 	logger.info("Comparing song collections")
 	to_upload = sorted(