+++ src/google_music_scripts/commands/upload.py
@@ -5,12 +5,12 @@ import google_music
 from logzero import logger
 
 from google_music_scripts.__about__ import __title__, __version__
-from google_music_scripts.cli import CustomPath, default_to_cwd, split_filter_strings
+from google_music_scripts.cli import CONTEXT_SETTINGS, CustomPath, default_to_cwd, parse_filters
 from google_music_scripts.config import configure_logging
 from google_music_scripts.core import get_local_songs, upload_songs
 
 
-@click.command()
+@click.command(context_settings=CONTEXT_SETTINGS)
 @click.version_option(__version__, '-V', '--version', prog_name=__title__, message="%(prog)s %(version)s")
 @click.option('-l', '--log', is_flag=True, default=False, help="Log to file.")
 @click.option('-v', '--verbose', count=True)
@@ -32,22 +32,6 @@ from google_music_scripts.core import get_local_songs, upload_songs
 	'--max-depth', metavar='DEPTH', type=int,
 	help="Set maximum depth of recursion when scanning for local files.\nDefault is infinite recursion."
 )
-@click.option(
-	'-f', '--include-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-fa', '--all-includes', is_flag=True, default=False,
-	help="Songs must match all include filter criteria to be included."
-)
-@click.option(
-	'-F', '--exclude-filter', metavar='FILTER', multiple=True, callback=split_filter_strings,
-	help="Metadata filters to match Google songs.\nSongs can match any filter criteria."
-)
-@click.option(
-	'-Fa', '--all-excludes', is_flag=True, default=False,
-	help="Songs must match all exclude filter criteria to be included."
-)
 @click.option('--delete-on-success', is_flag=True, default=False, help="Delete successfully uploaded local files.")
 @click.option(
 	'--transcode-lossless/--no-transcode-lossless', is_flag=True, default=True,
@@ -57,10 +41,14 @@ from google_music_scripts.core import get_local_songs, upload_songs
 	'--transcode-lossy/--no-transcode-lossy', is_flag=True, default=True,
 	help="Transcode non-MP3 lossy files to MP3 for upload."
 )
+@click.option(
+	'-f', '--filters', metavar='FILTER', multiple=True,
+	callback=parse_filters, help="Metadata filters."
+)
 @click.argument('input-paths', nargs=-1, type=CustomPath(resolve_path=True), callback=default_to_cwd)
 def upload(
-	log, verbose, quiet, dry_run, username, uploader_id, no_recursion, max_depth, delete_on_success,
-	transcode_lossless, transcode_lossy, include_filter, all_includes, exclude_filter, all_excludes, input_paths):
+	log, verbose, quiet, dry_run, username, uploader_id, no_recursion, max_depth,
+	delete_on_success, transcode_lossless, transcode_lossy, filters, input_paths):
 	"""Upload songs to a Google Music library."""
 
 	configure_logging(verbose - quiet, log_to_file=log)
@@ -76,12 +64,7 @@ def upload(
 	elif max_depth is None:
 		max_depth = float('inf')
 
-	to_upload = get_local_songs(
-		input_paths, max_depth=max_depth,
-		include_filters=include_filter, all_includes=all_includes,
-		exclude_filters=exclude_filter, all_excludes=all_excludes
-	)
-
+	to_upload = get_local_songs(input_paths, filters=filters, max_depth=max_depth)
 	to_upload.sort()
 
 	if not to_upload: